backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
python-multipart
pydantic
hf_xet
python-dotenv
openai
